Revamp of getdata to enable multiple line correctly. Validated by MainTest

+++ Source/main.h
@@ -20,12 +20,12 @@ void print_parse_error(std::string err);
 void print_readme();
 bool is_valid_name(std::string in);
 void checkformat(ifstream &file, unsigned int * axis1, unsigned int * axis2);
-SingleVector * getdata(ifstream &myfile, unsigned int axis1, unsigned int axis2);
+DoubleVector * getdata(ifstream &myfile, unsigned int axis1, unsigned int axis2);
 template <typename T>
 T StringToNumber(const string &in);
 template <typename T>
 string NumberToString(const T in);
-SingleVector * Read_data(const string filename);
+DoubleVector * Read_data(const string filename);
 std::string toString(complex in);
 
 #endif /* MAIN_H */
